+++ carina-api/src/test/java/com/qaprosoft/apitools/validation/mock/method/MockServer.java
@@ -1,4 +1,4 @@
-package com.qaprosoft.apitools.validation.method;
+package com.qaprosoft.apitools.validation.mock.method;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
