+++ carina-api/src/test/resources/api.properties
@@ -1,3 +1,3 @@
-PutDocMethod=PUT:{base_url}
-GetDocMethod=GET:${base_url}/docs
-GetUsersMethod=GET:${base_url}/users
+PutDocMethod=PUT:${base_url}
+NoContentTypeMethod=GET:${base_url}/mock1
+XmlContentTypeMethod=GET:${base_url}/mock2
