+++ carina-api/src/test/resources/config.properties
@@ -1,3 +1,3 @@
 env=DEMO
-DEMO.mockServer.port=8888
-DEMO.api_url=http://localhost:8888
+DEMO.mockServer.port=15100
+DEMO.api_url=http://localhost:15100
